requirements.txt
PyYAML>=6.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
cx-Oracle>=8.3.0
pyodbc>=4.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
Flask>=3.0.0
